+++ contracts/oracles/adaptors/chainlink/ChainlinkAdaptor.sol
@@ -148,7 +148,7 @@ contract ChainlinkAdaptor is BaseOracleAdaptor {
         );
 
         result.hadError = _verifyData(adjustedPrice, updatedAt, c.heartbeat);
-        result.price = uint240(adjustedPrice);
+        result.price = adjustedPrice;
     }
 
     /// @notice Wipes `asset` pricing configurations from this adaptor.
