+++ contracts/oracles/adaptors/redstone/RedstoneClassicAdaptor.sol
@@ -153,7 +153,7 @@ contract RedstoneClassicAdaptor is BaseOracleAdaptor {
         );
 
         result.hadError = _verifyData(adjustedPrice, updatedAt, c.heartbeat);
-        result.price = uint240(adjustedPrice);
+        result.price = adjustedPrice;
     }
 
     /// @notice Wipes `asset` pricing configurations from this adaptor.
